backend/extensions/k8s/tools/deployments.py
"""
Placeholder tools for K8s extension
These will be implemented based on the patterns established in kubectl.py and pods.py
"""

from goose.toolkit.base import Tool
from typing import Dict, Any, Optional

class ClusterInfoTool(Tool):
    def __init__(self, mcp_client=None):
        super().__init__(name="cluster_info", description="Get cluster information and health status")
        self.mcp_client = mcp_client
    
    async def execute(self) -> Dict[str, Any]:
        return {"success": True, "cluster": "info_placeholder"}

class DeploymentsTool(Tool):
    def __init__(self, default_namespace="default"):
        super().__init__(name="get_deployments", description="List and manage deployments")
        self.default_namespace = default_namespace
    
    async def execute(self, namespace: Optional[str] = None) -> Dict[str, Any]:
        return {"success": True, "deployments": []}

class ServicesTool(Tool):
    def __init__(self, default_namespace="default"):
        super().__init__(name="get_services", description="List services and endpoints")
        self.default_namespace = default_namespace
    
    async def execute(self, namespace: Optional[str] = None) -> Dict[str, Any]:
        return {"success": True, "services": []}

class NodesTool(Tool):
    def __init__(self, mcp_client=None):
        super().__init__(name="get_nodes", description="Get node status and information")
        self.mcp_client = mcp_client
    
    async def execute(self) -> Dict[str, Any]:
        return {"success": True, "nodes": []}
