backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Goose AI Agent Framework
goose-ai>=0.9.0

# Anthropic Claude API (for Goose provider)
anthropic==0.7.8

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Environment
python-dotenv==1.0.0

# CORS
fastapi-cors==0.0.6

# Logging
structlog==23.2.0

# JSON
orjson==3.9.10

# Kubernetes client
kubernetes>=29.0.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1
